feat(data-table): implement selection column logic with support for single selection mode and tree structure
